pyproject.toml
[project]
name = "embedding-models-benchmark"
version = "0.1.0"
description = "Benchmarking embedding models on BEIR retrieval tasks with hard negative analysis"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.50.0",
    "cohere>=5.11.0",
    "google-genai>=1.0.0",
    "python-dotenv>=1.0.1",
    "pyarrow>=17.0.0",
    "tqdm>=4.66.0",
    "numpy>=2.0.0",
    "matplotlib>=3.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/benchmark"]
